// src/lib/supabase/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
      listing_images: {
        Row: {
          created_at: string | null
          id: string
          listing_id: string
          order: number
          url: string
        }
        Insert: {
          created_at?: string | null
          id?: string
          listing_id: string
          order: number
          url: string
        }
        Update: {
          created_at?: string | null
          id?: string
          listing_id?: string
          order?: number
          url?: string
        }
        Relationships: [
          {
            foreignKeyName: "listing_images_listing_id_fkey"
            columns: ["listing_id"]
            isOneToOne: false
            referencedRelation: "listings"
            referencedColumns: ["id"]
          },
        ]
      }
      listings: {
        Row: {
          bathroom_count: number
          category: string
          created_at: string | null
          description: string
          guest_count: number
          id: string
          location_value: string
          price: number
          room_count: number
          title: string
          user_id: string
        }
        Insert: {
          bathroom_count: number
          category: string
          created_at?: string | null
          description: string
          guest_count: number
          id?: string
          location_value: string
          price: number
          room_count: number
          title: string
          user_id?: string
        }
        Update: {
          bathroom_count?: number
          category?: string
          created_at?: string | null
          description?: string
          guest_count?: number
          id?: string
          location_value?: string
          price?: number
          room_count?: number
          title?: string
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof Database },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof Database },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never